+++ build-logic/src/main/kotlin/Environment.kt
@@ -106,7 +106,7 @@ object Artifacts {
         platform = Java,
         groupId = "de.mannodermaus.gradle.plugins",
         artifactId = "android-junit5",
-        currentVersion = "2.0.0-SNAPSHOT",
+        currentVersion = "2.0.0",
         latestStableVersion = "1.14.0.0",
         description = "Unit Testing with the JUnit Framework for Android."
     )
@@ -116,8 +116,8 @@ object Artifacts {
      */
     object Instrumentation {
         const val groupId = "de.mannodermaus.junit5"
-        private const val currentVersion = "2.0.0"
-        private const val latestStableVersion = "1.9.0"
+        private const val currentVersion = "2.0.1-SNAPSHOT"
+        private const val latestStableVersion = "2.0.0"
 
         val Core = Deployed(
             platform = Android(minSdk = 19),